+++ src/utils/createOptionsList.js
@@ -1,0 +1,4 @@
+export const createOptionsList = (array, formatMessage) => array.map(option => ({
+  value: option.value,
+  label: formatMessage({ id: option.label }),
+}));