+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 * Refactor `bigtest/mirage` with `miragejs`.
 * MARC Bib field mapping profile: details for Update-Overrides on Create/Edit screen (UIDATIMP-631)
 * MARC Bib field mapping profile: details for Update-Overrides on View screen (UIDATIMP-632)
+* Change "Check in/out note" value to "Check in/out" for items (UIDATIMP-679)
 
 ### Bugs fixed:
 * Fix rendering qualifier sections with old data in match profiles details (UIDATIMP-481)